moved EmptyState folder and its content present inside ui/components to ui/components/shared/
